requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
numpy==2.0.2
openai==1.82.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==10.4.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruff==0.11.11
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tomlkit==0.12.0
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==12.0
zipp==3.21.0
